+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 3.14.1
 * Wooden trapdoors are now 33% less hard to break
 * Fixed blocking explosion with shields dealing no knockback
 * Fixed experimental flying block feature not making blocks fly